Rename file {specialist_document, section}_edition
