+++ app/models/section_edition.rb
