+++ app/repositories/repository_registry.rb
@@ -1,13 +1,12 @@
 require "specialist_document_repository"
 require "manual_repository"
-require "specialist_document_edition"
+require "section_edition"
 require "marshallers/document_association_marshaller"
 require "marshallers/manual_publish_task_association_marshaller"
 require "manual_publish_task"
 require "manual_with_publish_tasks"
 require "manual"
 require "manual_record"
-require "specialist_document_edition"
 require "specialist_document_repository"
 
 class RepositoryRegistry
