+++ spec/models/section_edition_spec.rb
